requirements.txt
python-dotenv==1.0.0
pydantic==2.5.2
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
asyncpg==0.29.0
motor==3.3.1
redis==5.0.1
tensorflow==2.16.2
keras==3.7.0
numpy==1.24.3
pandas==2.1.3
scikit-learn==1.3.2
requests==2.31.0
aiohttp==3.9.1
prometheus-client==0.19.0
python-json-logger==2.0.7 